Finish addrspace and thread management

+++ kernel/src/arch/aarch64/mm_type.rs
@@ -1,7 +1,7 @@
-pub const PAGE_SIZE:usize=1<<12;
-pub const PAGE_SIZE_BITS:usize=12;
-pub const VA_WIDTH:usize=48;
-pub const PA_WIDTH:usize=48;
-pub const KERNEL_BASE:usize=0xffff_0000_0000_0000;
+pub const PAGE_SIZE:u64=1<<12;
+pub const PAGE_SIZE_BITS:u64=12;
+pub const VA_WIDTH:u64=48;
+pub const PA_WIDTH:u64=48;
+pub const KERNEL_BASE:u64=0xffff_0000_0000_0000;
